+++ backend/src/main/java/todoktodok/backend/notification/application/service/command/NotificationCommandService.java
@@ -1,6 +1,5 @@
 package todoktodok.backend.notification.application.service.command;
 
-import java.util.List;
 import java.util.NoSuchElementException;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -8,7 +7,6 @@ import org.springframework.transaction.annotation.Transactional;
 import todoktodok.backend.member.domain.Member;
 import todoktodok.backend.member.domain.repository.MemberRepository;
 import todoktodok.backend.notification.application.dto.response.NotificationCheckResponse;
-import todoktodok.backend.notification.application.dto.response.NotificationResponse;
 import todoktodok.backend.notification.domain.Notification;
 import todoktodok.backend.notification.domain.repository.NotificationRepository;
 import todoktodok.backend.notification.exception.NotificationForbiddenException;
@@ -37,16 +35,6 @@ public class NotificationCommandService {
                 .build();
 
         notificationRepository.save(notification);
-    }
-
-    public List<NotificationResponse> getNotifications(final Long memberId) {
-        final Member recipient = findMember(memberId);
-        final Long notReadCount = notificationRepository.countNotificationByRecipientAndIsReadFalse(recipient);
-        final List<Notification> notifications = notificationRepository.findNotificationsByRecipient(recipient);
-
-        return notifications.stream()
-                .map(notification -> new NotificationResponse(notReadCount, notification))
-                .toList();
     }
 
     public void markNotificationAsRead(
