+++ backend/src/main/java/todoktodok/backend/notification/application/service/query/NotificationQueryService.java
@@ -1,0 +1,39 @@
+package todoktodok.backend.notification.application.service.query;
+
+import java.util.List;
+import java.util.NoSuchElementException;
+import lombok.AllArgsConstructor;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+import todoktodok.backend.member.domain.Member;
+import todoktodok.backend.member.domain.repository.MemberRepository;
+import todoktodok.backend.notification.application.dto.response.NotificationResponse;
+import todoktodok.backend.notification.domain.Notification;
+import todoktodok.backend.notification.domain.repository.NotificationRepository;
+
+@Service
+@Transactional(readOnly = true)
+@AllArgsConstructor
+public class NotificationQueryService {
+
+    private final NotificationRepository notificationRepository;
+    private final MemberRepository memberRepository;
+
+    public List<NotificationResponse> getNotifications(final Long memberId) {
+        final Member recipient = findMember(memberId);
+        final Long notReadCount = notificationRepository.countNotificationByRecipientAndIsReadFalse(recipient);
+        final List<Notification> notifications = notificationRepository.findNotificationsByRecipient(recipient);
+
+        return notifications.stream()
+                .map(notification -> new NotificationResponse(notReadCount, notification))
+                .toList();
+    }
+
+    private Member findMember(final Long memberId) {
+        return memberRepository.findByIdAndDeletedAtIsNull(memberId)
+                .orElseThrow(() -> new NoSuchElementException(
+                                String.format("해당하는 회원을 찾을 수 없습니다 : recipientId = %d", memberId)
+                        )
+                );
+    }
+}
